pyproject.toml
[project]
name = "fast-annotate"
version = "0.1.0"
description = "A fastHTML image annotator - Simple, keyboard-driven image annotation tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-fasthtml>=0.4.0",
    "pyyaml>=6.0",
    "starlette>=0.37.0",
]
